requirements.txt
pandas~=2.1.1
python-dotenv==1.0.0
Requests==2.31.0